Add multiprocessing support for oqmd
Add multiprocessing to the oqmd converter. Achieves approximately 50%
reduction in conversion time. More testing is necessary to optimize the
ratio of processors to writers. It is observed that the conversion
process slows as the conversion continues, down to a low of half-speed
by the end. This may be due to the increasing size of the feedstock
file, which means that there exist practical limits on feedstock size
for reasonable conversion times.