Rearrange repo in preparation for packaging
This commit breaks all of the converters (need to fix imports) and most
of the other code (need to make package for installation), too.
On the bright side, this commit also brings the repo into the
appropriate shape for packaging, and adds both the toolbox (with
utilities) and the refinery (placeholder).