Remove offset, test with longer timeout
Remove the offset argument from searches. Globus Search cannot guarantee that results are returned in the same order for the same query, so using an offset could easily lead to missed or duplicated results. Additionally, Globus Search will not process more than 10k results, ever, so the offset cannot be used to get, for example, the 10,001st result. If scrolling is enabled in the future, the offset will be brought back. In the meantime, the official method of retrieving the Nth result in Search is to search with limit=N and discard results 1 to N-1.
Also, increase the timeout for Travis testing to the maximum value.