Update SearchClient, start anonymous login
Toolbox has updated to the latest globus_sdk, where the SearchClient is now maintained. The SearchClient has also been updated, and both Forge and Query have been updated to match the new behavior. The only user-facing change is that the search index can be specified when searching, instead of only when instantiating a Forge object. Also moving the examples into the docs dir, to keep the Jupyte notebooks all in a separate space.